requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
ydata-profiling>=4.0.0
matplotlib>=3.5.0
weasyprint>=60.0
fpdf2>=2.0.0
